Move devices proxy to own module
